finsihed tests and javdocs
just need to finsih Intersection test, read me, and examples
