+++ docs/allclasses-frame.html
@@ -13,7 +13,7 @@
 <ul>
 <li><a href="kchandra423/kColliders/KCollider.html" title="class in kchandra423.kColliders" target="classFrame">KCollider</a></li>
 <li><a href="kchandra423/kColliders/KColliderContainer.html" title="class in kchandra423.kColliders" target="classFrame">KColliderContainer</a></li>
-<li><a href="kchandra423/kImages/KImage.html" title="class in kchandra423.kImages" target="classFrame">KImage</a></li>
+<li><a href="kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages" target="classFrame"><span class="interfaceName">KImage</span></a></li>
 <li><a href="kchandra423/kImages/KImageBuilder.html" title="class in kchandra423.kImages" target="classFrame">KImageBuilder</a></li>
 <li><a href="kchandra423/movableKImages/Movable.html" title="interface in kchandra423.movableKImages" target="classFrame"><span class="interfaceName">Movable</span></a></li>
 <li><a href="kchandra423/movableKImages/MovableKImage.html" title="class in kchandra423.movableKImages" target="classFrame">MovableKImage</a></li>
