+++ docs/allclasses-noframe.html
@@ -13,7 +13,7 @@
 <ul>
 <li><a href="kchandra423/kColliders/KCollider.html" title="class in kchandra423.kColliders">KCollider</a></li>
 <li><a href="kchandra423/kColliders/KColliderContainer.html" title="class in kchandra423.kColliders">KColliderContainer</a></li>
-<li><a href="kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a></li>
+<li><a href="kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages"><span class="interfaceName">KImage</span></a></li>
 <li><a href="kchandra423/kImages/KImageBuilder.html" title="class in kchandra423.kImages">KImageBuilder</a></li>
 <li><a href="kchandra423/movableKImages/Movable.html" title="interface in kchandra423.movableKImages"><span class="interfaceName">Movable</span></a></li>
 <li><a href="kchandra423/movableKImages/MovableKImage.html" title="class in kchandra423.movableKImages">MovableKImage</a></li>
