+++ docs/index-all.html
@@ -74,7 +74,7 @@
 <dl>
 <dt><span class="memberNameLink"><a href="kchandra423/kColliders/KCollider.html#copy--">copy()</a></span> - Method in class kchandra423.kColliders.<a href="kchandra423/kColliders/KCollider.html" title="class in kchandra423.kColliders">KCollider</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#copy--">copy()</a></span> - Method in class kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a></dt>
+<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#copy--">copy()</a></span> - Method in interface kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a></dt>
 <dd>
 <div class="block">Returns a copy of this image</div>
 </dd>
@@ -88,7 +88,7 @@
 </a>
 <h2 class="title">D</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#draw-processing.core.PApplet-float-float-">draw(PApplet, float, float)</a></span> - Method in class kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a></dt>
+<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#draw-processing.core.PApplet-float-float-">draw(PApplet, float, float)</a></span> - Method in interface kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a></dt>
 <dd>
 <div class="block">Draws this image at the given location on the given PApplet.</div>
 </dd>
@@ -115,7 +115,7 @@
 <div class="block">Returns a rectangle representing the boundaries of this area as the most
  tight rectangle possible</div>
 </dd>
-<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#getHeight--">getHeight()</a></span> - Method in class kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a></dt>
+<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#getHeight--">getHeight()</a></span> - Method in interface kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="kchandra423/kImages/Scalable.html#getHeight--">getHeight()</a></span> - Method in interface kchandra423.kImages.<a href="kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages">Scalable</a></dt>
 <dd>
@@ -123,7 +123,7 @@
 </dd>
 <dt><span class="memberNameLink"><a href="kchandra423/movableKImages/MovableKImage.html#getHeight--">getHeight()</a></span> - Method in class kchandra423.movableKImages.<a href="kchandra423/movableKImages/MovableKImage.html" title="class in kchandra423.movableKImages">MovableKImage</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#getImage--">getImage()</a></span> - Method in class kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a></dt>
+<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#getImage--">getImage()</a></span> - Method in interface kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a></dt>
 <dd>
 <div class="block">Gets the base Image used by this object</div>
 </dd>
@@ -143,7 +143,7 @@
 <dd>
 <div class="block">Returns the <code>PConstants</code> of this Image mode.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#getWidth--">getWidth()</a></span> - Method in class kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a></dt>
+<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#getWidth--">getWidth()</a></span> - Method in interface kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="kchandra423/kImages/Scalable.html#getWidth--">getWidth()</a></span> - Method in interface kchandra423.kImages.<a href="kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages">Scalable</a></dt>
 <dd>
@@ -219,12 +219,10 @@
 <dd>
 <div class="block">Creates a new KCollider Container with each KCollider being associated with a name in this hashmap</div>
 </dd>
-<dt><a href="kchandra423/kImages/KImage.html" title="class in kchandra423.kImages"><span class="typeNameLink">KImage</span></a> - Class in <a href="kchandra423/kImages/package-summary.html">kchandra423.kImages</a></dt>
+<dt><a href="kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages"><span class="typeNameLink">KImage</span></a> - Interface in <a href="kchandra423/kImages/package-summary.html">kchandra423.kImages</a></dt>
 <dd>
 <div class="block">A basic class for representing an Image.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="kchandra423/kImages/KImage.html#KImage--">KImage()</a></span> - Constructor for class kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a></dt>
-<dd>&nbsp;</dd>
 <dt><a href="kchandra423/kImages/KImageBuilder.html" title="class in kchandra423.kImages"><span class="typeNameLink">KImageBuilder</span></a> - Class in <a href="kchandra423/kImages/package-summary.html">kchandra423.kImages</a></dt>
 <dd>
 <div class="block">Defines a way to create KImages based on their path name, as they cannot be directly instantiated.</div>
