+++ docs/kchandra423/kColliders/KCollider.html
@@ -181,7 +181,7 @@ extends <a href="../../kchandra423/movableKImages/MovableKImage.html" title="cla
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
 <dd><a href="../../kchandra423/movableKImages/MovableKImage.html" title="class in kchandra423.movableKImages"><code>MovableKImage</code></a>, 
-<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages"><code>KImage</code></a>, 
+<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages"><code>KImage</code></a>, 
 <code>Area</code>, 
 <a href="../../kchandra423/kColliders/KColliderContainer.html" title="class in kchandra423.kColliders"><code>KColliderContainer</code></a></dd>
 </dl>
@@ -218,12 +218,12 @@ extends <a href="../../kchandra423/movableKImages/MovableKImage.html" title="cla
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/kColliders/KCollider.html#KCollider-kchandra423.kImages.KImage-">KCollider</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;k)</code>
+<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/kColliders/KCollider.html#KCollider-kchandra423.kImages.KImage-">KCollider</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;k)</code>
 <div class="block">Constructs a new KCollider with the given KImage.</div>
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/kColliders/KCollider.html#KCollider-kchandra423.kImages.KImage-float-float-float-boolean-kchandra423.movableKImages.MovableKImage.ImageMode-">KCollider</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;k,
+<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/kColliders/KCollider.html#KCollider-kchandra423.kImages.KImage-float-float-float-boolean-kchandra423.movableKImages.MovableKImage.ImageMode-">KCollider</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;k,
          float&nbsp;x,
          float&nbsp;y,
          float&nbsp;angle,
@@ -233,7 +233,7 @@ extends <a href="../../kchandra423/movableKImages/MovableKImage.html" title="cla
 </td>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/kColliders/KCollider.html#KCollider-kchandra423.kImages.KImage-float-float-float-boolean-kchandra423.movableKImages.MovableKImage.ImageMode-int-">KCollider</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;k,
+<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/kColliders/KCollider.html#KCollider-kchandra423.kImages.KImage-float-float-float-boolean-kchandra423.movableKImages.MovableKImage.ImageMode-int-">KCollider</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;k,
          float&nbsp;x,
          float&nbsp;y,
          float&nbsp;angle,
@@ -244,7 +244,7 @@ extends <a href="../../kchandra423/movableKImages/MovableKImage.html" title="cla
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/kColliders/KCollider.html#KCollider-kchandra423.kImages.KImage-int-">KCollider</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;k,
+<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/kColliders/KCollider.html#KCollider-kchandra423.kImages.KImage-int-">KCollider</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;k,
          int&nbsp;density)</code>
 <div class="block">Constructs a new KCollider with the given KImage at the given density.</div>
 </td>
@@ -396,7 +396,7 @@ extends <a href="../../kchandra423/movableKImages/MovableKImage.html" title="cla
 <ul class="blockList">
 <li class="blockList">
 <h4>KCollider</h4>
-<pre>public&nbsp;KCollider(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;k)</pre>
+<pre>public&nbsp;KCollider(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;k)</pre>
 <div class="block">Constructs a new KCollider with the given KImage. Default area density will be used.
  Starts at 0,0 with an angle of 0, unreflected and a corner <code>MovableKImage.ImageMode</code></div>
 <dl>
@@ -411,7 +411,7 @@ extends <a href="../../kchandra423/movableKImages/MovableKImage.html" title="cla
 <ul class="blockList">
 <li class="blockList">
 <h4>KCollider</h4>
-<pre>public&nbsp;KCollider(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;k,
+<pre>public&nbsp;KCollider(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;k,
                  int&nbsp;density)</pre>
 <div class="block">Constructs a new KCollider with the given KImage at the given density.
  Starts at 0,0 with an angle of 0, unreflected and a corner <code>MovableKImage.ImageMode</code></div>
@@ -428,7 +428,7 @@ extends <a href="../../kchandra423/movableKImages/MovableKImage.html" title="cla
 <ul class="blockList">
 <li class="blockList">
 <h4>KCollider</h4>
-<pre>public&nbsp;KCollider(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;k,
+<pre>public&nbsp;KCollider(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;k,
                  float&nbsp;x,
                  float&nbsp;y,
                  float&nbsp;angle,
@@ -453,7 +453,7 @@ extends <a href="../../kchandra423/movableKImages/MovableKImage.html" title="cla
 <ul class="blockListLast">
 <li class="blockList">
 <h4>KCollider</h4>
-<pre>public&nbsp;KCollider(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;k,
+<pre>public&nbsp;KCollider(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;k,
                  float&nbsp;x,
                  float&nbsp;y,
                  float&nbsp;angle,
