+++ docs/kchandra423/kImages/KImageBuilder.html
@@ -47,7 +47,7 @@ var activeTableTab = "activeTableTab";
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li><a href="../../kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
@@ -152,7 +152,7 @@ extends java.lang.Object</pre>
 <th class="colLast" scope="col">Method and Description</th>
 </tr>
 <tr id="i0" class="altColor">
-<td class="colFirst"><code>static <a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a></code></td>
+<td class="colFirst"><code>static <a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a></code></td>
 <td class="colLast"><code><span class="memberNameLink"><a href="../../kchandra423/kImages/KImageBuilder.html#getKImage-java.lang.String-">getKImage</a></span>(java.lang.String&nbsp;pathName)</code>
 <div class="block">Creates a KImage from its pathname.</div>
 </td>
@@ -202,7 +202,7 @@ extends java.lang.Object</pre>
 <ul class="blockListLast">
 <li class="blockList">
 <h4>getKImage</h4>
-<pre>public static&nbsp;<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;getKImage(java.lang.String&nbsp;pathName)
+<pre>public static&nbsp;<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;getKImage(java.lang.String&nbsp;pathName)
                         throws java.io.IOException</pre>
 <div class="block">Creates a KImage from its pathname. This library supports all Images supported by Java Advanced Imaging. While
  the specific file types supported vary drastically based on OS and OS version, but on my m1 macbook air I've found it
@@ -260,7 +260,7 @@ extends java.lang.Object</pre>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li><a href="../../kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
