+++ docs/kchandra423/kImages/Scalable.html
@@ -98,8 +98,12 @@ var activeTableTab = "activeTableTab";
 <ul class="blockList">
 <li class="blockList">
 <dl>
+<dt>All Known Subinterfaces:</dt>
+<dd><a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a></dd>
+</dl>
+<dl>
 <dt>All Known Implementing Classes:</dt>
-<dd><a href="../../kchandra423/kColliders/KCollider.html" title="class in kchandra423.kColliders">KCollider</a>, <a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>, <a href="../../kchandra423/movableKImages/MovableKImage.html" title="class in kchandra423.movableKImages">MovableKImage</a></dd>
+<dd><a href="../../kchandra423/kColliders/KCollider.html" title="class in kchandra423.kColliders">KCollider</a>, <a href="../../kchandra423/movableKImages/MovableKImage.html" title="class in kchandra423.movableKImages">MovableKImage</a></dd>
 </dl>
 <hr>
 <br>
