+++ docs/kchandra423/kImages/package-frame.html
@@ -12,11 +12,11 @@
 <div class="indexContainer">
 <h2 title="Interfaces">Interfaces</h2>
 <ul title="Interfaces">
+<li><a href="KImage.html" title="interface in kchandra423.kImages" target="classFrame"><span class="interfaceName">KImage</span></a></li>
 <li><a href="Scalable.html" title="interface in kchandra423.kImages" target="classFrame"><span class="interfaceName">Scalable</span></a></li>
 </ul>
 <h2 title="Classes">Classes</h2>
 <ul title="Classes">
-<li><a href="KImage.html" title="class in kchandra423.kImages" target="classFrame">KImage</a></li>
 <li><a href="KImageBuilder.html" title="class in kchandra423.kImages" target="classFrame">KImageBuilder</a></li>
 </ul>
 </div>
