+++ docs/kchandra423/kImages/package-summary.html
@@ -85,6 +85,12 @@
 </tr>
 <tbody>
 <tr class="altColor">
+<td class="colFirst"><a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a></td>
+<td class="colLast">
+<div class="block">A basic class for representing an Image.</div>
+</td>
+</tr>
+<tr class="rowColor">
 <td class="colFirst"><a href="../../kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages">Scalable</a></td>
 <td class="colLast">
 <div class="block">Defines a set of methods that any Scalable image should have; scaling and resizing, as well as the ability to get
@@ -103,12 +109,6 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colFirst"><a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a></td>
-<td class="colLast">
-<div class="block">A basic class for representing an Image.</div>
-</td>
-</tr>
-<tr class="rowColor">
 <td class="colFirst"><a href="../../kchandra423/kImages/KImageBuilder.html" title="class in kchandra423.kImages">KImageBuilder</a></td>
 <td class="colLast">
 <div class="block">Defines a way to create KImages based on their path name, as they cannot be directly instantiated.</div>
