+++ docs/kchandra423/kImages/package-tree.html
@@ -79,14 +79,17 @@
 <ul>
 <li type="circle">java.lang.Object
 <ul>
-<li type="circle">kchandra423.kImages.<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages"><span class="typeNameLink">KImage</span></a> (implements kchandra423.kImages.<a href="../../kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages">Scalable</a>)</li>
 <li type="circle">kchandra423.kImages.<a href="../../kchandra423/kImages/KImageBuilder.html" title="class in kchandra423.kImages"><span class="typeNameLink">KImageBuilder</span></a></li>
 </ul>
 </li>
 </ul>
 <h2 title="Interface Hierarchy">Interface Hierarchy</h2>
 <ul>
-<li type="circle">kchandra423.kImages.<a href="../../kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages"><span class="typeNameLink">Scalable</span></a></li>
+<li type="circle">kchandra423.kImages.<a href="../../kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages"><span class="typeNameLink">Scalable</span></a>
+<ul>
+<li type="circle">kchandra423.kImages.<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages"><span class="typeNameLink">KImage</span></a></li>
+</ul>
+</li>
 </ul>
 </div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
