+++ docs/kchandra423/movableKImages/MovableKImage.html
@@ -162,12 +162,12 @@ implements <a href="../../kchandra423/kImages/Scalable.html" title="interface in
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/movableKImages/MovableKImage.html#MovableKImage-kchandra423.kImages.KImage-">MovableKImage</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;kImage)</code>
+<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/movableKImages/MovableKImage.html#MovableKImage-kchandra423.kImages.KImage-">MovableKImage</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;kImage)</code>
 <div class="block">Creates a new Movable KImage at 0,0 with angle 0, not reflected, and an ImageMode of corner</div>
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/movableKImages/MovableKImage.html#MovableKImage-kchandra423.kImages.KImage-float-float-float-boolean-kchandra423.movableKImages.MovableKImage.ImageMode-">MovableKImage</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;kImage,
+<td class="colOne"><code><span class="memberNameLink"><a href="../../kchandra423/movableKImages/MovableKImage.html#MovableKImage-kchandra423.kImages.KImage-float-float-float-boolean-kchandra423.movableKImages.MovableKImage.ImageMode-">MovableKImage</a></span>(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;kImage,
              float&nbsp;x,
              float&nbsp;y,
              float&nbsp;angle,
@@ -216,7 +216,7 @@ implements <a href="../../kchandra423/kImages/Scalable.html" title="interface in
 </td>
 </tr>
 <tr id="i4" class="altColor">
-<td class="colFirst"><code><a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a></code></td>
+<td class="colFirst"><code><a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a></code></td>
 <td class="colLast"><code><span class="memberNameLink"><a href="../../kchandra423/movableKImages/MovableKImage.html#getImage--">getImage</a></span>()</code>
 <div class="block">Returns the base KImage used by this object</div>
 </td>
@@ -321,7 +321,7 @@ implements <a href="../../kchandra423/kImages/Scalable.html" title="interface in
 <ul class="blockList">
 <li class="blockList">
 <h4>MovableKImage</h4>
-<pre>public&nbsp;MovableKImage(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;kImage)</pre>
+<pre>public&nbsp;MovableKImage(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;kImage)</pre>
 <div class="block">Creates a new Movable KImage at 0,0 with angle 0, not reflected, and an ImageMode of corner</div>
 <dl>
 <dt><span class="paramLabel">Parameters:</span></dt>
@@ -335,7 +335,7 @@ implements <a href="../../kchandra423/kImages/Scalable.html" title="interface in
 <ul class="blockListLast">
 <li class="blockList">
 <h4>MovableKImage</h4>
-<pre>public&nbsp;MovableKImage(<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;kImage,
+<pre>public&nbsp;MovableKImage(<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;kImage,
                      float&nbsp;x,
                      float&nbsp;y,
                      float&nbsp;angle,
@@ -438,7 +438,7 @@ implements <a href="../../kchandra423/kImages/Scalable.html" title="interface in
 <ul class="blockList">
 <li class="blockList">
 <h4>getImage</h4>
-<pre>public&nbsp;<a href="../../kchandra423/kImages/KImage.html" title="class in kchandra423.kImages">KImage</a>&nbsp;getImage()</pre>
+<pre>public&nbsp;<a href="../../kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages">KImage</a>&nbsp;getImage()</pre>
 <div class="block">Returns the base KImage used by this object</div>
 <dl>
 <dt><span class="returnLabel">Returns:</span></dt>
