+++ docs/overview-tree.html
@@ -82,7 +82,6 @@
 <li type="circle">java.lang.Object
 <ul>
 <li type="circle">kchandra423.kColliders.<a href="kchandra423/kColliders/KColliderContainer.html" title="class in kchandra423.kColliders"><span class="typeNameLink">KColliderContainer</span></a></li>
-<li type="circle">kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="class in kchandra423.kImages"><span class="typeNameLink">KImage</span></a> (implements kchandra423.kImages.<a href="kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages">Scalable</a>)</li>
 <li type="circle">kchandra423.kImages.<a href="kchandra423/kImages/KImageBuilder.html" title="class in kchandra423.kImages"><span class="typeNameLink">KImageBuilder</span></a></li>
 <li type="circle">kchandra423.movableKImages.<a href="kchandra423/movableKImages/MovableKImage.html" title="class in kchandra423.movableKImages"><span class="typeNameLink">MovableKImage</span></a> (implements kchandra423.kImages.<a href="kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages">Scalable</a>)
 <ul>
@@ -95,7 +94,11 @@
 <h2 title="Interface Hierarchy">Interface Hierarchy</h2>
 <ul>
 <li type="circle">kchandra423.movableKImages.<a href="kchandra423/movableKImages/Movable.html" title="interface in kchandra423.movableKImages"><span class="typeNameLink">Movable</span></a></li>
-<li type="circle">kchandra423.kImages.<a href="kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages"><span class="typeNameLink">Scalable</span></a></li>
+<li type="circle">kchandra423.kImages.<a href="kchandra423/kImages/Scalable.html" title="interface in kchandra423.kImages"><span class="typeNameLink">Scalable</span></a>
+<ul>
+<li type="circle">kchandra423.kImages.<a href="kchandra423/kImages/KImage.html" title="interface in kchandra423.kImages"><span class="typeNameLink">KImage</span></a></li>
+</ul>
+</li>
 </ul>
 <h2 title="Enum Hierarchy">Enum Hierarchy</h2>
 <ul>
